refactor: move aws-markerplace and copilotkit routes to App router
