+++ keep-ui/app/api/copilotkit/route.ts
@@ -1,0 +1,24 @@
+import {
+  CopilotRuntime,
+  OpenAIAdapter,
+  copilotRuntimeNextJSAppRouterEndpoint,
+} from "@copilotkit/runtime";
+import OpenAI from "openai";
+import { NextRequest } from "next/server";
+
+const openai = new OpenAI({
+  organization: process.env.OPEN_AI_ORGANIZATION_ID,
+  apiKey: process.env.OPEN_AI_API_KEY,
+});
+const serviceAdapter = new OpenAIAdapter({ openai });
+const runtime = new CopilotRuntime();
+
+export const POST = async (req: NextRequest) => {
+  const { handleRequest } = copilotRuntimeNextJSAppRouterEndpoint({
+    runtime,
+    serviceAdapter,
+    endpoint: "/api/copilotkit",
+  });
+
+  return handleRequest(req);
+};
