--- keep-ui/pages/api/copilotkit.ts
@@ -1,28 +1,0 @@
-import { NextApiRequest, NextApiResponse } from "next";
-import {
-  CopilotRuntime,
-  OpenAIAdapter,
-  copilotRuntimeNextJSPagesRouterEndpoint,
-} from "@copilotkit/runtime";
-import OpenAI from "openai";
-
-const openai = new OpenAI({
-  organization: process.env.OPEN_AI_ORGANIZATION_ID,
-  apiKey: process.env.OPEN_AI_API_KEY,
-});
-
-export default async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse
-) {
-  const serviceAdapter = new OpenAIAdapter({ openai });
-  const runtime = new CopilotRuntime();
-
-  const handleRequest = copilotRuntimeNextJSPagesRouterEndpoint({
-    endpoint: "/api/copilotkit",
-    runtime,
-    serviceAdapter,
-  });
-
-  return await handleRequest(req, res);
-}
